feat: add file validations

+++ frontend/src/components/FileCard.tsx
@@ -3,7 +3,7 @@ import {
     AiOutlineDelete as CloseIcon,
 } from "solid-icons/ai";
 import { Component } from "solid-js";
-import { getCourseDetails } from "../utils/courseDetails";
+import { validateFilename } from "../utils/validateFilename";
 
 type Props = {
     file: File;
@@ -11,30 +11,32 @@ type Props = {
 };
 
 export const FileCard: Component<Props> = ({ file, removeFile }) => {
-    const courseDetails = getCourseDetails(file.name);
-    return (
-        <div class="file-card">
-            <PDFIcon size="4.5rem" />
-            <div class="file-data">
-                <h4 class="file-name">{file.name}</h4>
-                <div class="course-name">
-                    {`${courseDetails.course_code} - ${courseDetails.course_name}`}
+    const courseDetails = validateFilename(file.name);
+    if (courseDetails)
+        return (
+            <div class="file-card">
+                <PDFIcon size="4.5rem" />
+                <div class="file-data">
+                    <h4 class="file-name">{file.name}</h4>
+                    <div class="course-name">
+                        {`${courseDetails.course_code} - ${courseDetails.course_name}`}
+                    </div>
+                    <div class="pills">
+                        {courseDetails.year && (
+                            <div class="pill">{courseDetails.year}</div>
+                        )}
+                        {courseDetails.exam && (
+                            <div class="pill">{courseDetails.exam}</div>
+                        )}
+                        {courseDetails.semester && (
+                            <div class="pill">{courseDetails.semester}</div>
+                        )}
+                    </div>
                 </div>
-                <div class="pills">
-                    {courseDetails.year && (
-                        <div class="pill">{courseDetails.year}</div>
-                    )}
-                    {courseDetails.exam && (
-                        <div class="pill">{courseDetails.exam}</div>
-                    )}
-                    {courseDetails.semester && (
-                        <div class="pill">{courseDetails.semester}</div>
-                    )}
-                </div>
+                <button onClick={() => removeFile(file.name)} class="close-btn">
+                    <CloseIcon size="2rem" />
+                </button>
             </div>
-            <button onClick={() => removeFile(file.name)} class="close-btn">
-                <CloseIcon size="2rem" />
-            </button>
-        </div>
-    );
+        );
+    else return <></>;
 };
